refactor: tweak
