+++ web-to-node-handler-abort-signal/README.md
@@ -1,6 +1,6 @@
 https://github.com/remix-run/remix/issues/9438
 
 ```sh
-SERVER_ENTRY=/src/hattip pnpm dev
 SERVER_ENTRY=/src/hono pnpm dev
+SERVER_ENTRY=/src/hattip pnpm dev
 ```
