+++ web-to-node-handler-abort-signal/src/hono.ts
@@ -1,27 +1,4 @@
 import { getRequestListener } from "@hono/node-server";
-import { Hono } from "hono";
+import { handler } from "./handler";
 
-const app = new Hono();
-
-app.get("*", () => {
-	let cancelled = false;
-
-	const stream = new ReadableStream<string>({
-		async start(controller) {
-			for (let i = 0; !cancelled; i++) {
-				console.log({ i });
-				controller.enqueue(`i = ${i}\n`);
-				await new Promise((resolve) => setTimeout(resolve, 1000));
-			}
-			controller.close();
-		},
-		cancel() {
-			console.log("cancel!");
-			cancelled = true;
-		},
-	});
-
-	return new Response(stream.pipeThrough(new TextEncoderStream()));
-});
-
-export default getRequestListener(app.fetch);
+export default getRequestListener(handler);
